requirements.txt
python-telegram-bot==20.7
feedparser==6.0.10
requests==2.31.0
yfinance==0.2.32
python-dotenv==1.0.0
pytz==2023.3
python-dateutil==2.8.2
